refactor: keep slack profile lookup in enrichment job
